Add Auth

+++ display-form.php
@@ -28,17 +28,26 @@ require('header.php');
 
     //start the table
     echo '<table class="table table-bordered table-striped text-center"><thead><th>Name</th><th>Address</th><th>Phone</th><th>Gender</th>
-<th>Product Name</th><th>Product Price</th><th>Actions</th></thead><tbody>';
+<th>Product Name</th><th>Product Price</th>';
+
+    // Check Auth
+    if(isset($_SESSION['userId'])){
+        echo '<th>Actions</th>' ;
+    }
+    echo'</thead><tbody>';
 
     //loop and print the data
     foreach($accInfo as $ac){
         echo'<tr><td>'. $ac['name'] .'</td><td>' . $ac['address']. '</td><td>' . $ac['phone'] . '</td><td>' . $ac['gender'] . '</td><td>' .
             $ac['productName'] . '</td><td>' . $ac['productPrice'] . '</td>' ;
 
-    echo "<td><a class=\"btn btn-sm btn-info\" href=\"eBuyForm.php?accountId={$ac['accountId']}\">Edit</a>
-                <a href=\"delete.php?accountId={$ac['accountId']}\"
-                class=\"btn btn-sm btn-danger confirmation\">Delete</a></td>";
-    echo '</tr>';
+        if(isset($_SESSION['userId'])){
+            echo "<td><a class=\"btn btn-sm btn-info\" href=\"eBuyForm.php?accountId={$ac['accountId']}\">Edit</a>
+                    <a href=\"delete.php?accountId={$ac['accountId']}\"
+                    class=\"btn btn-sm btn-danger confirmation\">Delete</a></td>";
+        }
+
+        echo '</tr>';
     }
 
     //close the table
